+++ src/main/java/ladder/domain/player/Player.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.player;
 
 public class Player extends Name {
