+++ src/test/java/ladder/domain/NameTest.java
@@ -1,5 +1,6 @@
 package ladder.domain;
 
+import ladder.domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
