+++ src/test/java/ladder/domain/ResultTest.java
@@ -1,5 +1,6 @@
 package ladder.domain;
 
+import ladder.domain.result.Result;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
